refactor: 웹 로또 관련 로직을 LottoWebController에 모두 반영
